Feat: add reset button
There are a few changes related to how we find the sdl2 libraries on the
system. Before this commit we used find_package but that had some
troubles finding the SDL_ttf library. So, now we have switched to
PkgConfig to find the sdl2 libraries, which seems to be working fine for
now.
A res directory has also been added which contains all the resource that
the game needs. Currently we only have a font. In future we should but
the tile and tile maps in there.