Fix: check if the tile is moveable in moveTile
Before this commit we were not checking the type of the tile i.e.
whether it was moveable or not. After writing some tests I realised that
there was a segfault when moving a tile of type brick out of bounds.
After this commit there shouldn't be any segfault for moving a tile out
of bounds that is either moveable or non-moveable.