Feat: add timer feature to the game
This commits add a timer feature, since right now there's only a single
level in the game so the game exits as soon as the time is up. In future
there should be a life feature so when the timer is over the life
reduces.
The timer is an interface and the GameTimer implements the timer
interface in order to provide all the time related details (it is built
on top of SDL_GetTicks function). The timer allows testable code (we can
create a spy timer to test the game behaviour).