Test: add test for game timer logic
This commit adds a spy timer that implements the timer interface to
inject the dependency. The benefit of creating a spy timer is that we
don't slow down our tests by waiting for the actual timer to finish.
Instead we can set the ticks as per our need to test the game timer
logic.